Update to 5.2-SNAPSHOT
