+++ open-metadata-implementation/adapters/open-connectors/repository-services-connectors/open-metadata-collection-store-connectors/docs/README.md
@@ -43,7 +43,7 @@ following:
 
 ```
 <properties>
-    <open-metadata.version>5.1-SNAPSHOT</open-metadata.version>
+    <open-metadata.version>5.2-SNAPSHOT</open-metadata.version>
 </properties>
 <dependencies>
     <dependency>
