+++ open-metadata-implementation/platform-chassis/platform-chassis-spring/src/main/java/org/odpi/openmetadata/platformchassis/springboot/OMAGServerPlatform.java
@@ -53,7 +53,7 @@ import org.apache.lucene.queryparser.classic.QueryParserBase;
 @OpenAPIDefinition(
         info = @Info(
                 title = "Egeria's Open Metadata and Governance (OMAG) Server Platform",
-                version = "5.1-SNAPSHOT",
+                version = "5.2-SNAPSHOT",
                 description = "The OMAG Server Platform provides a runtime process and platform for Open Metadata and Governance (OMAG) Services.\n" +
                         "\n" +
                         "The OMAG services are configured and activated in OMAG Servers using the Administration Services.\n" +
