+++ open-metadata-implementation/platform-services/platform-services-server/src/main/java/org/odpi/openmetadata/platformservices/server/OMAGServerPlatformOriginServices.java
@@ -8,7 +8,7 @@ package org.odpi.openmetadata.platformservices.server;
  */
 public class OMAGServerPlatformOriginServices
 {
-    final String   implementationOrigin = "Egeria OMAG Server Platform (version 5.1-SNAPSHOT)\n";
+    final String   implementationOrigin = "Egeria OMAG Server Platform (version 5.2-SNAPSHOT)\n";
 
     /**
      * Return the origin of this server platform implementation.
