+++ open-metadata-resources/open-metadata-archives/content-pack-helpers/src/main/java/org/odpi/openmetadata/samples/archiveutilities/EgeriaBaseArchiveWriter.java
@@ -35,7 +35,7 @@ public abstract class EgeriaBaseArchiveWriter extends OMRSArchiveWriter
     /*
      * Specific values for initializing TypeDefs
      */
-    protected static final String versionName   = "5.1-SNAPSHOT";
+    protected static final String versionName   = "5.2-SNAPSHOT";
 
     protected       OMRSArchiveBuilder      archiveBuilder;
     protected       GovernanceArchiveHelper archiveHelper;
