+++ open-metadata-resources/open-metadata-deployment/docker/configure/Dockerfile
@@ -2,7 +2,7 @@
 # Copyright Contributors to the Egeria project
 
 FROM docker.io/library/alpine:3.20.0
-ARG version=5.1-SNAPSHOT
+ARG version=5.2-SNAPSHOT
 ARG VCS_REF=unknown
 ARG VCS_ORIGIN=unknown
 ARG BUILD_TIME=unknown
