+++ open-metadata-resources/open-metadata-deployment/docker/platform/README.md
@@ -107,7 +107,7 @@ With that in place we can now run our docker image, this time making use of the 
 $ docker run -p 9443:9443 -v source=egeria-data,target=/deployments/platform/data odpi/egeria-platform:latest 
 /usr/local/s2i/run: line 15: /opt/jboss/container/maven/default//scl-enable-maven: No such file or directory
 Starting the Java application using /opt/jboss/container/java/run/run-java.sh ...
-INFO exec  java -XX:+UseParallelOldGC -XX:MinHeapFreeRatio=10 -XX:MaxHeapFreeRatio=20 -XX:GCTimeRatio=4 -XX:AdaptiveSizePolicyWeight=90 -XX:MaxMetaspaceSize=100m -XX:+ExitOnOutOfMemoryError -XX:MaxMetaspaceSize=1g -cp "." -jar /deployments/platform/omag-server-platform-5.1-SNAPSHOT.jar
+INFO exec  java -XX:+UseParallelOldGC -XX:MinHeapFreeRatio=10 -XX:MaxHeapFreeRatio=20 -XX:GCTimeRatio=4 -XX:AdaptiveSizePolicyWeight=90 -XX:MaxMetaspaceSize=100m -XX:+ExitOnOutOfMemoryError -XX:MaxMetaspaceSize=1g -cp "." -jar /deployments/platform/omag-server-platform-5.2-SNAPSHOT.jar
  ODPi Egeria
     ____   __  ___ ___    ______   _____                                 ____   _         _     ___
    / __ \ /  |/  //   |  / ____/  / ___/ ___   ____ _   __ ___   ____   / _  \ / / __    / /  / _ /__   ____ _  _
@@ -176,7 +176,7 @@ you may need to create another job to retrieve the required libraries or content
 #### Adding to the image through a docker build
 
 Use the egeria image as a base, for example begin your custom Dockerfile with
-`FROM odpi/egeria-platform:5.1-SNAPSHOT`
+`FROM odpi/egeria-platform:5.2-SNAPSHOT`
 
 Then add in the files you need, as well as customize the LOADER_PATH variable ie
 `COPY myextralib.jar /deployments/platform/extra`
